+++ src/main/java/com/ChaveMestra/Application/service/CategoriaService.java
@@ -1,6 +1,8 @@
 package com.ChaveMestra.Application.service;
 import com.ChaveMestra.Application.dto.CategoriaRequest;
 import com.ChaveMestra.Application.dto.CategoriaResponse;
+import com.ChaveMestra.Application.exception.BusinessException;
+import com.ChaveMestra.Application.exception.ResourceNotFoundException;
 import com.ChaveMestra.Application.mapper.CategoriaMapper;
 import com.ChaveMestra.Application.model.Categoria;
 import com.ChaveMestra.Application.repository.ProdutoRepository;
@@ -44,8 +46,11 @@ public class CategoriaService {
 
     @Transactional
     public void excluir(Integer id) {
+        if (!categoriaRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Categoria não encontrada");
+        }
         if (produtoRepository.existsByCategoriaId(id)) {
-            throw new RuntimeException("Nao foi possivel excluir");
+            throw new BusinessException("Não foi possível excluir a categoria com produtos vinculados");
         }
         categoriaRepository.deleteById(id);
     }
@@ -53,7 +58,7 @@ public class CategoriaService {
     @Transactional(readOnly = true)
     public CategoriaResponse buscarPorId(Integer id) {
         Optional<Categoria> objetoBuscado = categoriaRepository.findById(id);
-        Categoria categoria = objetoBuscado.orElseThrow(() -> new RuntimeException("Categoria não encontrada"));
+        Categoria categoria = objetoBuscado.orElseThrow(() -> new ResourceNotFoundException("Categoria não encontrada"));
         return categoriaMapper.toResponse(categoria);
     }
 
@@ -61,13 +66,12 @@ public class CategoriaService {
         @Transactional
         public CategoriaResponse atualizar(Integer id, CategoriaRequest dados) {
             Categoria categoria = categoriaRepository.findById(id)
-                    .orElseThrow(() -> new RuntimeException("Categoria nao existente"));
+                    .orElseThrow(() -> new ResourceNotFoundException("Categoria não encontrada"));
             categoria.setNome(dados.nome());
             categoria.setDescricao(dados.descricao());
             categoria = categoriaRepository.save(categoria);
             return categoriaMapper.toResponse(categoria);
         }
     }
-
